feat(images): add & view
- Implementado endpoints para subir y hacer fecth de imagenes
- Modificados JSPs para cargar la imagen desde el atributo de Producto
- Modificados JSs para cargar la imagen desde el atributo de Producto

+++ src/main/java/es/laboticademar/webstore/controllers/restControllers/ProductRestController.java
@@ -47,8 +47,8 @@ import lombok.RequiredArgsConstructor;
         @RequestParam(required = false) Boolean activo,
         @RequestParam(required = false) Boolean stock,
         @RequestParam(required = false) Boolean conDescuento, // <-- AÑADIR ESTA LÍNEA
-        @RequestParam(required = false) BigDecimal precioMin,
-        @RequestParam(required = false) BigDecimal precioMax
+        @RequestParam(required = false) Float precioMin,
+        @RequestParam(required = false) Float precioMax
     ) {
 
         Page<ProductoDTO> productPage = productService.getAllProducts(
